redas-data-clean/requirements.txt
requests
beautifulsoup4=4.6.0
kafka-python=1.4.7